Compose data finalizer instead of inheriting it.
Added the concept of a data finalizer which adds assembly to the
the shellcode to initialize the offset register (GET_PC) and the
data. Since all finalize methods are now more or less the same, move
that up to BaseEnvironment and make PREAMBLE a property of that class.

The default data finalizer (mutable segment) is implemented in three
forms: gnu_as_mutable_data_finalizer, nasm_mutable_data_finalizer and
nasm_null_safe_mutable_data_finalizer. The first one is a generic
one for use with GNU as, the last two are for nasm (and thus x86).